requirements.txt
Flask==3.0.3
Flask-WTF==1.2.1
Werkzeug==3.0.3
huggingface_hub==0.27.0
openpyxl==3.1.2
python-dotenv==1.0.1
APScheduler==3.10.4
qrcode[pil]==7.4.2
gunicorn==21.2.0
# Optional: Turso/libsql client (requires Rust toolchain on Windows)
# libsql-experimental==0.0.55
